Move LangChain tests into main package and eliminate workspace
