+++ aws-distro-opentelemetry-node-autoinstrumentation/package.json
@@ -34,8 +34,10 @@
     "prewatch": "npm run precompile",
     "prepublishOnly": "npm run compile",
     "tdd": "yarn test -- --watch-extensions ts --watch",
-    "test": "nyc ts-mocha --timeout 10000 -p tsconfig.json --require '@opentelemetry/contrib-test-utils' 'test/**/*.ts'",
-    "test:coverage": "nyc --check-coverage --functions 95 --lines 95 ts-mocha --timeout 10000 -p tsconfig.json --require '@opentelemetry/contrib-test-utils' 'test/**/*.ts'",
+    "test": "npm run test:unit && npm run test:langchain",
+    "test:unit": "nyc ts-mocha --timeout 10000 -p tsconfig.json --require '@opentelemetry/contrib-test-utils' --ignore 'test/instrumentation-langchain/**' 'test/**/*.ts'",
+    "test:langchain": "nyc ts-mocha --timeout 10000 -p tsconfig.json --require '@opentelemetry/contrib-test-utils' 'test/instrumentation-langchain/**/*.ts'",
+    "test:coverage": "nyc --check-coverage --functions 95 --lines 95 ts-mocha --timeout 10000 -p tsconfig.json --require '@opentelemetry/contrib-test-utils' --ignore 'test/instrumentation-langchain/**' 'test/**/*.ts'",
     "watch": "tsc -w"
   },
   "nyc": {
@@ -84,23 +86,35 @@
     "@aws-sdk/client-sfn": "3.982.0",
     "@aws-sdk/client-sts": "3.982.0",
     "@aws-sdk/client-sns": "3.982.0",
+    "@langchain/anthropic": "^1.3.26",
+    "@langchain/aws": "^1.3.5",
+    "@langchain/cohere": "^1.0.4",
+    "@langchain/core": "^1.0.0",
+    "@langchain/deepseek": "^1.0.24",
+    "@langchain/google-genai": "^2.1.27",
+    "@langchain/groq": "^1.2.0",
+    "@langchain/mistralai": "^1.0.7",
+    "@langchain/openai": "^1.4.4",
+    "@langchain/xai": "^1.3.16",
     "@opentelemetry/contrib-test-utils": "^0.61.0",
+    "@smithy/node-http-handler": "^4.0.0",
     "@smithy/protocol-http": "^5.0.1",
     "@smithy/signature-v4": "^5.0.1",
     "@types/mocha": "7.0.2",
     "@types/node": "18.6.5",
     "@types/proxyquire": "^1.3.31",
-    "@langchain/core": "^0.3.0",
     "@types/sinon": "10.0.18",
     "expect": "29.2.0",
     "mocha": "7.2.0",
     "nock": "13.2.1",
+    "node-fetch": "^2.7.0",
     "nyc": "15.1.0",
     "proxyquire": "^2.1.3",
     "rimraf": "5.0.5",
     "sinon": "15.2.0",
     "ts-mocha": "10.0.0",
-    "typescript": "5.0.4"
+    "typescript": "5.0.4",
+    "zod": "^3.24.0"
   },
   "dependencies": {
     "@aws-sdk/client-cloudwatch-logs": "3.982.0",
