--- test/instrumentation-langchain/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@aws/test-instrumentation-langchain",
-  "version": "0.0.0",
-  "private": true,
-  "description": "Integration tests for LangChain instrumentation with real provider libraries",
-  "scripts": {
-    "test": "nyc ts-mocha --timeout 10000 -p tsconfig.json --require '@opentelemetry/contrib-test-utils' 'test/**/*.ts'",
-    "tdd": "npm run test -- --watch-extensions ts --watch"
-  },
-  "devDependencies": {
-    "@aws-sdk/client-bedrock-runtime": "3.982.0",
-    "@aws/aws-distro-opentelemetry-node-autoinstrumentation": "file:../../aws-distro-opentelemetry-node-autoinstrumentation",
-    "@langchain/anthropic": "^1.3.26",
-    "@langchain/aws": "^1.3.5",
-    "@langchain/cohere": "^1.0.4",
-    "@langchain/core": "^1.0.0",
-    "@langchain/deepseek": "^1.0.24",
-    "@langchain/google-genai": "^2.1.27",
-    "@langchain/groq": "^1.2.0",
-    "@langchain/mistralai": "^1.0.7",
-    "@langchain/openai": "^1.4.4",
-    "@langchain/xai": "^1.3.16",
-    "@opentelemetry/api": "1.9.1",
-    "@opentelemetry/contrib-test-utils": "^0.61.0",
-    "@opentelemetry/sdk-trace-base": "2.6.1",
-    "@opentelemetry/semantic-conventions": "1.40.0",
-    "@smithy/node-http-handler": "^4.0.0",
-    "@types/mocha": "7.0.2",
-    "@types/node": "18.6.5",
-    "expect": "29.2.0",
-    "mocha": "7.2.0",
-    "nock": "13.2.1",
-    "nyc": "15.1.0",
-    "ts-mocha": "10.0.0",
-    "typescript": "5.0.4"
-  }
-}
