--- test/instrumentation-langchain/tsconfig.json
@@ -1,33 +1,0 @@
-{
-  "compilerOptions": {
-    "rootDir": ".",
-    "outDir": "build",
-    "skipLibCheck": true,
-    "allowUnreachableCode": false,
-    "allowUnusedLabels": false,
-    "declaration": true,
-    "declarationMap": true,
-    "forceConsistentCasingInFileNames": true,
-    "inlineSources": true,
-    "module": "commonjs",
-    "noEmitOnError": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitOverride": true,
-    "noImplicitReturns": true,
-    "noUnusedLocals": true,
-    "pretty": true,
-    "sourceMap": true,
-    "strict": true,
-    "strictNullChecks": true,
-    "target": "es2017",
-    "incremental": true,
-    "newLine": "LF",
-    "resolveJsonModule": true
-  },
-  "include": [
-    "test/**/*.ts"
-  ],
-  "exclude": [
-    "node_modules"
-  ]
-}
